move BTDownload component to control/
