+++ locales/zh-CN.json
@@ -305,6 +305,20 @@
   "WeChat QR Code Title": "请使用微信扫码",
   "Bind WeChat Text in WeChatBind": "确定使用以下帐号绑定Wisnuc吗？",
   "Bind WeChat Success": "成功绑定微信成功",
+ 
+  "==== control/BTDownload.js ====": "",
+  "BT Download Folder Name": "Download",
+  "Download Folder EEXIST Text": "无法创建目录‘Download‘，可能存在重名文件！",
+  "BT Start Failed": "无法启动BT下载",
+  "No BT Task Text": "请点击左上按钮添加新的下载任务",
+  "Add Torrent": "添加BT种子文件",
+  "Add Magnet": "添加磁力链接",
+  "Add Magnet Label": "请输入磁力链接地址",
+  "Delete Task Text": "确定要删除选中的任务吗？",
+  "Add Magnet Failed": "添加失败",
+  "Add Magnet Success": "添加成功",
+  "Finished": "已完成",
+  "Getting Info": "获取信息中",
 
   "==== device/DeviceInfo.jsx ====": "",
   "Modify Device Name Failed": "修改失败",
@@ -732,19 +746,5 @@
   "Read Upload Failed": "读取上传文件失败",
 
   "==== node/lib/newDownload.js ====": "",
-  "Read Download Failed": "读取下载目录失败",
-  
-  "==== download/BTDownload.js ====": "",
-  "BT Download Folder Name": "Download",
-  "Download Folder EEXIST Text": "无法创建目录‘Download‘，可能存在重名文件！",
-  "BT Start Failed": "无法启动BT下载",
-  "No BT Task Text": "请点击左上按钮添加新的下载任务",
-  "Add Torrent": "添加BT种子文件",
-  "Add Magnet": "添加磁力链接",
-  "Add Magnet Label": "请输入磁力链接地址",
-  "Delete Task Text": "确定要删除选中的任务吗？",
-  "Add Magnet Failed": "添加失败",
-  "Add Magnet Success": "添加成功",
-  "Finished": "已完成",
-  "Getting Info": "获取信息中"
+  "Read Download Failed": "读取下载目录失败"
 }
